백엔드 클라스, 메서드 이름 변경
