+++ miniproject/src/main/java/com/kh/miniproject/controller/ListController.java
@@ -5,7 +5,9 @@ import com.kh.miniproject.vo.ListVo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
@@ -15,15 +17,32 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ListController {
     private final ListService listService;
-    
-    @PostMapping("/search")
-    public ResponseEntity<List<ListVo>> getFiltered(@RequestBody ListVo listVo) {
-        log.warn("실행");
-        log.error(listVo.toString());
-        // 서비스 로직 호출 (필터링된 자동차 목록 반환)
-        List<ListVo> filteredCars = listService.getFiltered(listVo);
-        log.error(filteredCars.toString());
-        // 결과 반환
-        return ResponseEntity.ok(filteredCars);
+
+    @GetMapping("/companies")
+    public ResponseEntity<List<ListVo>> getManufacturer() {
+        log.warn("제조사 컨트롤러 시작");
+        List<ListVo> manufacturers = listService.getManufacturer();
+        log.warn("제조사 : {}", manufacturers);
+        return ResponseEntity.ok(manufacturers);
+    }
+
+    @GetMapping("/engines")
+    public ResponseEntity<List<ListVo>> getEngine() {
+        log.warn("엔진 컨트롤러 시작");
+        List<ListVo> engines = listService.getEngines();
+        log.warn("엔진 : {}", engines);
+        return ResponseEntity.ok(engines);
+    }
+
+    @GetMapping("/maxPrice")
+    public ResponseEntity<Integer> getMaxPrice() {
+        Integer maxPrice = listService.getMaxPrice();
+        return ResponseEntity.ok(maxPrice);
+    }
+
+    @GetMapping("/carClasses")
+    public ResponseEntity<List<ListVo>> getClassification() {
+        List<ListVo> classifications = listService.getClassification();
+        return ResponseEntity.ok(classifications);
     }
 }
