+++ miniproject/src/main/java/com/kh/miniproject/service/ListService.java
@@ -6,32 +6,34 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
-
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class ListService {
+
     private final ListRepository listRepository;
     
-    public List<ListVo> getFiltered(ListVo listVo) {
-        String carName = listVo.getCarName();
-        String manufacturer = listVo.getManufacturer();
-        Integer minPrice = listVo.getMinPrice();
-        Integer maxPrice = listVo.getMaxPrice();
-        String engineType = listVo.getEngineType();
-        String classification = listVo.getClassification();
-        String sortBy = listVo.getSortBy();
-        String sortType = listVo.getSortType();
-        log.error("service : carName = {}, manufacturer = {}, min / max = {} / {}, engine = {}, classification = {}, sort = {}, sortType = {}",
-                carName, manufacturer, minPrice, maxPrice, engineType, classification, sortBy, sortType);
-        
-        log.info("정렬 기준: {} 정렬 방향: {}", sortBy, sortType);
-        
-        // Repository 에서 직접 정렬된 데이터 가져오기
-        return listRepository.getFilter(carName, manufacturer, minPrice, maxPrice, engineType, classification, sortBy, sortType);
+    public List<ListVo> getManufacturer() {
+        log.warn("제조사 서비스 작동");
+        return listRepository.getManufacturer();
     }
 
+    public List<ListVo> getEngines() {
+        log.warn("엔진 서비스 작동");
+        return listRepository.getEngines();
+    }
+
+    public Integer getMaxPrice() {
+        log.warn("최대 가격 서비스 작동");
+        return listRepository.getMaxPrice();
+    }
+
+    public List<ListVo> getClassification() {
+        log.warn("차종 서비스 작동");
+        return listRepository.getCarClasses();
+    }
 
 }
+
+
