+++ miniproject/src/main/java/com/kh/miniproject/service/SortService.java
@@ -1,39 +1,36 @@
 package com.kh.miniproject.service;
 
 import com.kh.miniproject.repository.SortRepository;
-import com.kh.miniproject.vo.OptionVo;
+import com.kh.miniproject.vo.CategoryVo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class SortService {
-
     private final SortRepository sortRepository;
     
-    public List<OptionVo> getManufacturer() {
-        log.warn("제조사 서비스 작동");
-        return sortRepository.getManufacturer();
+    public List<CategoryVo> sortInfo(CategoryVo categoryVo) {
+        String carName = categoryVo.getCarName();
+        String manufacturer = categoryVo.getManufacturer();
+        Integer minPrice = categoryVo.getMinPrice();
+        Integer maxPrice = categoryVo.getMaxPrice();
+        String engineType = categoryVo.getEngineType();
+        String classification = categoryVo.getClassification();
+        String sortBy = categoryVo.getSortBy();
+        String sortType = categoryVo.getSortType();
+        log.error("service : carName = {}, manufacturer = {}, min / max = {} / {}, engine = {}, classification = {}, sort = {}, sortType = {}",
+                carName, manufacturer, minPrice, maxPrice, engineType, classification, sortBy, sortType);
+        
+        log.info("정렬 기준: {} 정렬 방향: {}", sortBy, sortType);
+        
+        // Repository 에서 직접 정렬된 데이터 가져오기
+        return sortRepository.sortInfo(carName, manufacturer, minPrice, maxPrice, engineType, classification, sortBy, sortType);
     }
 
-    public List<OptionVo> getEngines() {
-        log.warn("엔진 서비스 작동");
-        return sortRepository.getEngines();
-    }
-
-    public Integer getMaxPrice() {
-        log.warn("최대 가격 서비스 작동");
-        return sortRepository.getMaxPrice();
-    }
-
-    public List<OptionVo> getClassification() {
-        log.warn("차종 서비스 작동");
-        return sortRepository.getCarClasses();
-    }
 
 }
-
-
