+++ miniproject/src/main/java/com/kh/miniproject/vo/CategoryVo.java
@@ -1,0 +1,21 @@
+/* 정렬 메인페이지에서 차량 정렬을 위한 객체를 담는 VO */
+package com.kh.miniproject.vo;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class CategoryVo {
+    private int carNo;
+    private String carName;
+    private String manufacturer;
+    private Integer price;
+    private Integer minPrice;
+    private Integer maxPrice;
+    private String engineType;
+    private String classification;
+    private String sortBy;
+    private String sortType;
+}
