--- miniproject/src/main/java/com/kh/miniproject/vo/ExampleTableTest.java
@@ -1,14 +1,0 @@
-//package com.kh.miniproject.vo;
-//
-//import jakarta.persistence.*;
-//
-//@Entity
-//@Table(name = "EXAMPLE_TABLE_TEST")
-//public class ExampleTableTest {
-//    @Id
-//    @Column (name = "ID", nullable = false)
-//    private Long id;
-//
-//    @Column(name = "NAME", nullable = false)
-//    private String name;
-//}
