+++ miniproject/src/main/java/com/kh/miniproject/vo/ListVo.java
@@ -1,19 +1,14 @@
-/* 정렬 메인페이지에서 차량 정렬을 위한 객체를 담는 VO */
+/* 차량 필터시 사용되는 Vo */
 package com.kh.miniproject.vo;
 
 import lombok.*;
 
-@Data
+@Getter
+@Setter
+@AllArgsConstructor
 @NoArgsConstructor
 public class ListVo {
-    private int carNo;
-    private String carName;
-    private String manufacturer;
-    private Integer price;
-    private Integer minPrice;
-    private Integer maxPrice;
-    private String engineType;
-    private String classification;
-    private String sortBy;
-    private String sortType;
+	private String name;
+	private String image;
+	private String link;
 }
