--- miniproject/src/main/java/com/kh/miniproject/vo/OptionVo.java
@@ -1,14 +1,0 @@
-/* 차량 필터시 사용되는 Vo */
-package com.kh.miniproject.vo;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class OptionVo {
-	private String name;
-	private String image;
-	private String link;
-}
